Repurposed flf_pirates, added dvk05 (repurposed empa06).
Originally I was going to have a pirate fallout, but I decided to
instead just have some pirates attacking FLF ships for no real
reason (they're pirates, after all). The new filler mission,
"Alliance of Inconvenience", now unlocks the slightly altered
"Pirate Disturbance" mission, which is common, but not as common as
the others. It has a particular disadvantage compared to the others
in that it cannot give you reputation for the Pirate faction (makes
sense), but it does pay you and raise FLF reputation.
