+++ dat/faction.xml
@@ -183,6 +183,7 @@
   <enemies>
    <enemy>Dvaered</enemy>
    <enemy>Rogue FLF</enemy>
+   <enemy>Rogue Pirate</enemy>
   </enemies>
  </faction>
  <faction>
@@ -265,6 +266,18 @@
    <enemy>FLF</enemy>
    <enemy>Dvaered</enemy>
    <enemy>Empire</enemy>
+  </enemies>
+ </faction>
+ <faction>
+  <name>Rogue Pirate</name>
+  <display>Pirate</display>
+  <static />
+  <invisible />
+  <player>-100</player>
+  <colour>red</colour>
+  <standing>static</standing>
+  <enemies>
+   <enemy>FLF</enemy>
   </enemies>
  </faction>
  <faction>
